warning system rewrite

+++ database.py
@@ -1,5 +1,7 @@
 import os
 import pymongo
+import random
+import string
 from datetime import datetime
 from dotenv import load_dotenv
 
@@ -11,38 +13,33 @@ def current_time():
     return datetime.now().strftime('%d/%m/%Y %H:%M:%S')
 
 
+def random_id():
+    return ''.join(random.choice(string.ascii_letters + string.digits) for i in range(6))
+
+
 class Cosplayer:
     def __init__(self, user_id):
         self.user_id = user_id
 
-    # # warning system
     async def get_warnings(self):
-        return client['logs']['warnings'].find_one({'_id': self.user_id})
+        return list(client['logs']['warnings'].find({'user_id': self.user_id}))
 
     async def add_warning(self, warner: str, warning: str):
         try:
             cl = client['logs']['warnings']
-            try:
-                user_warnings = cl.find_one(({'_id': self.user_id}))['warnings']
 
-                # Check if lower warning number is available
-                warn_number = 0
-                for i in range(int(max(user_warnings))):
-                    try:
-                        user_warnings[str(i)]
-                    except KeyError:
-                        warn_number = i
+            rid = random_id()
+            while cl.find_one({'_id': rid}):
+                rid = random_id()
 
-                warn_number = int(max(user_warnings)) + 1 if warn_number == 0 else warn_number
+            cl.insert_one({
+                '_id': rid,
+                'user_id': self.user_id,
+                'warning': warning,
+                'warner': warner,
+                'time': current_time()
+            })
 
-                user_warnings[str(warn_number)] = {
-                    'warning': warning,
-                    'warner': warner,
-                    'time': current_time()
-                }
-            except (TypeError, ValueError):
-                user_warnings = {'1': {'warning': warning, 'warner': warner, 'time': current_time()}}
-            cl.update_one({'_id': self.user_id}, {'$set': {'warnings': user_warnings}}, upsert=True)
             return True
         except Exception as e:
             print(f'{current_time()} - Error: {e}')
@@ -50,22 +47,27 @@ class Cosplayer:
 
     async def remove_warning(self, warning_number):
         try:
-            cl = client['logs']['warnings']
-            try:
-                user_warnings = cl.find_one({'_id': self.user_id})['warnings']
-                user_warnings.pop(warning_number)
-                cl.update_one({'_id': self.user_id}, {'$set': {'warnings': user_warnings}}, upsert=True)
-                return True
-            except (KeyError, TypeError):
-                return False
+            client['logs']['warnings'].delete_one({'_id': warning_number, 'user_id': self.user_id})
+            return True
         except Exception as e:
             print(f'{current_time()} - Error: {e}')
             return 'error'
 
     async def reset_warnings(self):
         try:
-            client['logs']['warnings'].delete_one({'_id': self.user_id})
+            client['logs']['warnings'].delete_many({'user_id': self.user_id})
             return True
         except Exception as e:
             print(f'{current_time()} - Error: {e}')
             return 'error'
+
+
+import asyncio
+cosplayer = Cosplayer(298890523454734336)
+# print(asyncio.run(cosplayer.add_warning('database', 'test')))
+
+print(asyncio.run(cosplayer.reset_warnings()))
+print(asyncio.run(cosplayer.get_warnings()))
+
+# for warning in asyncio.run(cosplayer.get_warnings()):
+#     print(warning)
